+++ src/main/kotlin/vecharia/menu/PauseMenu.kt
@@ -4,7 +4,7 @@ import vecharia.Vecharia
 import vecharia.util.Tickable
 
 class PauseMenu(title: String) : Tickable {
-    override fun tick(game: Vecharia, frame: Int) {
+    override fun tick(game: Vecharia, frame: Long) {
         // todo implement
     }
 }
